Renames and deleted unnecesary dependencies
