+++ src/routes/safe/components/Apps/components/AppsList.tsx
@@ -14,7 +14,7 @@ import { SAFELIST_ADDRESS } from 'src/routes/routes'
 import { useAppList } from '../hooks/useAppList'
 import { SAFE_APP_FETCH_STATUS, SafeApp } from '../types'
 import AddAppForm from './AddAppForm'
-import { AppData } from '../api/fetchSafeAppsList'
+import { AppData } from '../api/config-service'
 
 const Wrapper = styled.div`
   height: 100%;
@@ -81,12 +81,12 @@ const AppContainer = styled.div`
 `
 
 const isAppLoading = (app: SafeApp) => SAFE_APP_FETCH_STATUS.LOADING === app.fetchStatus
-const isCustomApp = (appUrl: string, staticAppsList: AppData[]) => !staticAppsList.some(({ url }) => url === appUrl)
+const isCustomApp = (appUrl: string, appsList: AppData[]) => !appsList.some(({ url }) => url === appUrl)
 
 const AppsList = (): React.ReactElement => {
   const matchSafeWithAddress = useRouteMatch<{ safeAddress: string }>({ path: `${SAFELIST_ADDRESS}/:safeAddress` })
   const safeAddress = useSelector(safeParamAddressFromStateSelector)
-  const { appList, removeApp, staticAppsList, isLoading } = useAppList(true)
+  const { appList, removeApp, apiAppsList, isLoading } = useAppList()
   const [isAddAppModalOpen, setIsAddAppModalOpen] = useState<boolean>(false)
   const [appToRemove, setAppToRemove] = useState<SafeApp | null>(null)
 
@@ -120,7 +120,7 @@ const AppsList = (): React.ReactElement => {
                 <StyledLink key={a.url} to={`${matchSafeWithAddress?.url}/apps?appUrl=${encodeURI(a.url)}`}>
                   <AppCard isLoading={isAppLoading(a)} iconUrl={a.iconUrl} name={a.name} description={a.description} />
                 </StyledLink>
-                {isCustomApp(a.url, staticAppsList) && (
+                {isCustomApp(a.url, apiAppsList) && (
                   <IconBtn
                     title="Remove"
                     onClick={(e) => {
